MultiReceivers support for IRremote library
Unfortunately it has a dependency on my library CppList also which I use
in other projects. Because of this #include "CppList.h" needs to be
added into each example and project where IRremote needs to be used.
There must be a way to automatically include it together with usual
#include "IRremote.h", but I am not an experienced C++ coder to do this
in a proper way. I hope somebody else can do this.