Marantz-RC5 extension support (#1314)
* Marantz-RC5 extension support

Added support for Marantz extension of RC5 protocol. Marantz uses standard RC5x  (13 bit) as well as an own version that adds a pause (4x 889 us) after the address and add adds another word (the "extension") to the signal so the the extended message is 19 bit long.

* Send function for Marantz variant of RC5x

Added support for Marantz variant of RC5 protocol (sending only). Marantz uses standard RC5x (13 bit) as well as an own version that adds a pause (4x 889 us) after the address and add adds another word (the "command extension") to the signal so the extended message is 19 bit long.

---------

Co-authored-by: ArminJo <armin.arduino@gmail.com>